requirements.txt
dbus-python==1.2.18
font-roboto==0.0.1
fonts==0.0.3
Pillow==9.1.1
playsound==1.3.0
pulsectl==22.3.2
PyGObject==3.42.1
RPi.GPIO==0.7.1
ST7789==0.0.4